requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
scikit-learn==1.3.2
numpy==1.24.3
ollama==0.1.8
faster-whisper==0.10.0
pydantic==2.5.0
python-dotenv==1.0.0